+++ changelog.md
@@ -20,6 +20,7 @@ Check the command for more infos.
 * Snow golems can now strafe when attacking (disabled by default)
 * Fixed temporary cobwebs breaking any block in the cobweb's position
 * Fixed disabling "Melee Attacking" feature making melee attacks have no cooldown
+* Fixed fisher mobs canceling the fishing rod cast if they can't see the target
 
 ## 2.7.1
 * Villagers no longer use target changes